se modifico grafica

+++ app/src/main/java/com/example/christianfranco/basedatos/Status.java
@@ -105,10 +105,10 @@ public class Status extends AppCompatActivity {
 //seteo el color del background y el mensaje del dialogo
     public void color (){
         if (!(res.get(0).equals("Sin Registro"))){
-            if (Integer.valueOf(res.get(0)) > 200) {//si es mayor a 200 es DIABETES
+            if (Float.valueOf(res.get(0)) > 200) {//si es mayor a 200 es DIABETES
                 TG.setBackgroundResource(R.color.ROJO);
                 respuestaG="Diabetes";
-            } else if (144 < Integer.valueOf(res.get(0)) && Integer.valueOf(res.get(0)) < 149) {//entre 144 y 149 es PREDIABETES
+            } else if (144 < Float.valueOf(res.get(0)) && Integer.valueOf(res.get(0)) < 149) {//entre 144 y 149 es PREDIABETES
                 TG.setBackgroundResource(R.color.AMARILLO);
                 respuestaG="PRE-DIABETES";
             } else {//Normal
